Chore: Pin images, update dependencies and security audit (#695)
* Pin images, update deps, and adjust tests/UI

Pin base images and Composer in Dockerfiles, pin runtime images in docker-compose, and switch CI lint install to npm ci. Add a new GitHub Actions security workflow for dependency and container scanning. Bump several JS dev/runtime dependencies in package.json and adjust related code (react-day-picker upgrade reflected by classname change in calendar, replace initialFocus with autoFocus in date pickers). Update tests to be more robust: disable throttle middleware for email verification, set cache driver to array and fake queues/release locks in assistance tests, cast cached value to int in job test, manage cache driver in DataCite service tests, and switch to userEvent with async clicks in a vitest UI test.

* Add Playwright browser deps; update packages

Install Playwright browser runtime dependencies in Dockerfile.dev (fonts, X11, libGL, etc.) to support Pest Browser smoke tests. Update package.json and package-lock.json to use @playwright/test and playwright v1.59.1 so test tooling matches the runtime environment and avoids version mismatches during CI/dev runs.

* Adjust Trivy scan to report then fail on criticals

Update security workflow to separate reporting and failing behavior for Trivy scans. The SARIF step now uses exit-code '0' so results are uploaded (format: sarif), and a new Trivy step (format: table) is added to explicitly exit with '1' when CRITICAL or HIGH image vulnerabilities are found. Scanner, vuln-type, ignore-unfixed settings and action version are preserved.

* Remove build-time Node artifacts in Dockerfile

Cleanup Node build artifacts after running the frontend build to reduce runtime image size and remove leftover credentials/caches. The Dockerfile now runs npm build, removes public/hot, node_modules, npm cache directories, and package files (.npmrc, package.json, package-lock.json), and updates the comment to reflect the cleanup.

* Run Trivy via Docker image and add cache dir

Replace the aquasecurity/trivy-action steps with direct docker runs of aquasec/trivy:0.70.0. Add a .trivy-cache preparation step and mount the workspace and docker socket so Trivy can use a local cache (--cache-dir) and produce SARIF and table outputs. Also set TRIVY_DISABLE_VEX_NOTICE, switch to the newer --pkg-types flag, and preserve behavior to ignore unfixed vulnerabilities while failing the job on CRITICAL/HIGH findings (exit code 1). This avoids the action and gives more control over Trivy invocation and caching.

* Move Node.js install to build stage

Refactor Dockerfile into multi-stage build: introduce app-base (runtime), app-build (installs Node.js, runs npm build) and final app stage that copies built artifacts. This removes Node.js and package manifests from the runtime image, reducing image size and attack surface; NodeSource GPG key installation and apt-get nodejs are now confined to the build stage.

* Pin Trivy image and avoid SARIF upload from forks

Replace aquasec/trivy:0.70.0 tag with a sha256 digest for reproducible, immutable runner image references (updated both Trivy invocations). Also tighten the SARIF upload step condition to skip uploading results for pull requests from forked repos, preventing unintended uploads from external contributors.

+++ tests/pest/Feature/Jobs/DiscoverAssistantSuggestionsJobTest.php
@@ -296,7 +296,7 @@ describe('handle', function () {
         $job->handle($registrar);
 
         expect(Cache::tags($tags)->has($totalCountKey))->toBeTrue();
-        expect(Cache::tags($tags)->get($totalCountKey))->toBe(10);
+        expect((int) Cache::tags($tags)->get($totalCountKey))->toBe(10);
     });
 });
